requirements.txt
Django==4.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
python-dotenv==1.0.1
